Refactoring Provider Componenet
